Connect Lifetime
TCP connection lifetime is actually is a pain. Knowing if the client has disconnected, while supporting long polling are opposite goals essentially.

This patch adds a zero sized message the the message protocol and it's variants.

A client sending this to the server indicates that it is disconnecting.

The server sending this to the client is an attempt to determine if the long poll should continue.

Timeout is set at the generous 60 second mark. This should of course be a configuration in the future.